# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Work Order Services",
        "description": "Microservice for mapping work order to user",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user-workorders": {
            "get": {
                "summary": "Get Work Orders for a given user",
                "description": "Get Work Orders for a given user\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Work Order Mapping"
                ],
                "responses": {
                    "200": {
                        "description": "List of Work Orders",
                        "schema": {
                            "$ref": "#/definitions/response"
                        }
                    },
                    "500": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Associate Work Order to a user",
                "description": "Associate Work Order to a user \n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mapping",
                        "in": "body",
                        "description": "content",
                        "schema": {
                            "$ref": "#/definitions/workOrderMapping"
                        }
                    }
                ],
                "tags": [
                    "User Work Order Mapping"
                ],
                "responses": {
                    "200": {
                        "description": "Associated Work Order mapping entity",
                        "schema": {
                            "$ref": "#/definitions/response"
                        }
                    },
                    "500": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        },
        "response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "status": {
                    "type": "object"
                }
            }
        },
        "workOrderMapping": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                },
                "workOrderId": {
                    "type": "number"
                }
            }
        }
    }
}